initial version of the scope tab
