+++ src/client-scopes/messages.json
@@ -53,6 +53,7 @@
     "predefinedMappingDescription": "Choose one of the predefined mappings from this table",
     "mappingTable": "Table with predefined mapping",
     "roleGroup": "Use a realm role from:",
-    "clientGroup": "Use a client role from:"
+    "clientGroup": "Use a client role from:",
+    "scope": "Scope"
   }
 }
